Rename metadata => metadataManager
