+++ __tests__/unit/lib/metadata/metadataManager.test.js
@@ -1,5 +1,5 @@
 'use strict'
-const metadata = require('../../../../lib/metadata/metadata')
+const metadata = require('../../../../lib/metadata/metadataManager')
 
 describe(`test if metadata`, () => {
   test('when grouped per directoryName, have classes', () => {
