+++ lib/diffHandler.js
@@ -3,7 +3,7 @@ const path = require('path')
 const os = require('os')
 const childProcess = require('child_process')
 const TypeHandlerFactory = require('./service/typeHandlerFactory')
-const metadata = require('./metadata/metadata')('directoryName')
+const metadata = require('./metadata/metadataManager')('directoryName')
 
 const UTF8_ENCODING = 'utf8'
 const diffParams = ['diff', '--name-status', '--no-renames']
