--- lib/metadata/metadata.js
@@ -1,7 +1,0 @@
-'use strict'
-const describeMetadata = require('./v46.json')
-module.exports = grouping =>
-  describeMetadata.reduce((metadata, describe) => {
-    metadata[describe[grouping]] = describe
-    return metadata
-  }, {})
