+++ lib/packageConstructor.js
@@ -1,6 +1,6 @@
 'use strict'
 const xmlbuilder = require('xmlbuilder')
-const metadata = require('./metadata/metadata')('directoryName')
+const metadata = require('./metadata/metadataManager')('directoryName')
 const xmlConf = { indent: '    ', newline: '\n', pretty: true }
 
 module.exports = class PackageConstructor {
