+++ lib/service/inFolderHandler.js
@@ -1,5 +1,5 @@
 'use strict'
-const metadata = require('../metadata/metadata')('directoryName')
+const metadata = require('../metadata/metadataManager')('directoryName')
 const StandardHandler = require('./standardHandler')
 const path = require('path')
