+++ lib/service/standardHandler.js
@@ -1,5 +1,5 @@
 'use strict'
-const metadata = require('../metadata/metadata')('directoryName')
+const metadata = require('../metadata/metadataManager')('directoryName')
 const path = require('path')
 const fse = require('fs-extra')
