+++ lib/service/typeHandlerFactory.js
@@ -4,7 +4,7 @@ const InResource = require('./inResourceHandler')
 const Lightning = require('./lightningHandler')
 const InFolder = require('./inFolderHandler')
 const Standard = require('./standardHandler')
-const metadata = require('../metadata/metadata')('directoryName')
+const metadata = require('../metadata/metadataManager')('directoryName')
 const path = require('path')
 
 const classes = {
